+++ src/andrewtools/__init__.py
@@ -2,3 +2,4 @@
 #  - `from andrewtools.progress_bar import progress_bar` to
 #  - `from andrewtools import progress_bar`
 from .progress_bar import progress_bar
+from .andrewtimer import AndrewTimer