+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
 
 <!-- Placeholder For Next Release -->
-## v0.0.2 (09/15/2022)
+## v0.1.0 (09/19/2022)
 
 ### Features
 
+- Added `AndrewTimer` class, an easy-to-user execution timer
 - Shortened name of `print_progress_bar` to `progress_bar`
+
+### Test
+
+- Added tests for `AndrewTimer` and updated tests for `progress_bar`
 
 ## v0.0.1 (09/15/2022)